@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "firebase": "^9.17.1",
     "particle.js": "^1.0.7",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
+    "react": "16.8",
+    "react-dom": "16.8",
     "react-ga": "^2.5.3",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.3",
