@@ -1,30 +1,50 @@
-import React, { Component } from "react";
-import Demo from "./Demo";
-import ItemControl from "../../components/layout/ItemControl";
-import VideoBack from "../../components/layout/VideoBackGound";
-import { Consumer } from "../../context";
+import React, { useState, useEffect } from 'react';
+import Demo from './Demo';
+import ItemControl from '../../components/layout/ItemControl';
+import VideoBack from '../../components/layout/VideoBackGound';
+import { Consumer } from '../../context';
+import { useFirebase } from '../../firebase-context';
+import { collection, getDocs, orderBy, query } from 'firebase/firestore';
 
-class Demos extends Component {
-  state = {
-    video: "assets/videos/demo.mp4"
-  };
-  render() {
-    return (
-      <Consumer>
-        {value => {
-          const { demos, index, direction } = value;
-          return (
-            <section className="section demos">
-              <VideoBack video={this.state.video} />
+export default function Demos() {
+  const [video] = useState('assets/videos/demo.mp4');
+  const firebase = useFirebase();
+  const [demos, setDemos] = useState([]);
 
+  useEffect(() => {
+    async function getDemos() {
+      const querySnapshot = await getDocs(
+        query(collection(firebase.database(), 'demos'), orderBy('id', 'asc'))
+      );
+      const result = [];
+      querySnapshot.forEach((doc) => {
+        result.push(doc.data());
+      });
+
+      setDemos(result);
+    }
+
+    getDemos();
+  }, []);
+
+  return (
+    <Consumer>
+      {(value) => {
+        const { index, direction } = value;
+        return (
+          <section className="section demos">
+            <VideoBack video={video} />
+
+            {demos.length && (
               <Demo items={demos} index={index} direction={direction} />
-              <ItemControl name={"Demo"} length={demos.length} />
-            </section>
-          );
-        }}
-      </Consumer>
-    );
-  }
-}
+            )}
 
-export default Demos;
+            {demos.length && (
+              <ItemControl name={'Demo'} length={demos.length} />
+            )}
+          </section>
+        );
+      }}
+    </Consumer>
+  );
+}
