@@ -1,34 +1,49 @@
-import React, { Component } from "react";
-import Project from "./Project";
-import ItemControl from "../../components/layout/ItemControl";
-import VideoBack from "../../components/layout/VideoBackGound";
-import { Consumer } from "../../context";
+import React, { useState, useEffect } from 'react';
+import Project from './Project';
+import ItemControl from '../../components/layout/ItemControl';
+import VideoBack from '../../components/layout/VideoBackGound';
+import { Consumer } from '../../context';
+import { useFirebase } from '../../firebase-context';
+import { collection, orderBy, getDocs, query } from 'firebase/firestore';
 
-class Projects extends Component {
-  state = {
-    video: "assets/videos/project.mp4"
-  };
-  render() {
-    return (
-      <Consumer>
-        {value => {
-          const { projects, index, direction } = value;
-          return (
-            <section className="section projects">
-              <VideoBack video={this.state.video} />
-              <Project
-                items={projects}
-                index={index}
-                direction={direction}
-                transRef={this.transition}
-              />
-              <ItemControl length={projects.length} name={"Project"} />
-            </section>
-          );
-        }}
-      </Consumer>
-    );
-  }
+export default function Projects() {
+  const [video] = useState('assets/videos/project.mp4');
+  const firebase = useFirebase();
+  const [projects, setProjects] = useState([]);
+
+  useEffect(() => {
+    async function getProjects() {
+      const querySnapshot = await getDocs(
+        query(collection(firebase.database(), 'projects'), orderBy('id', 'asc'))
+      );
+      const result = [];
+      querySnapshot.forEach((doc) => {
+        result.push(doc.data());
+      });
+
+      setProjects(result);
+    }
+
+    getProjects();
+  }, []);
+
+  return (
+    <Consumer>
+      {(value) => {
+        const { index, direction } = value;
+        return (
+          <section className="section projects">
+            <VideoBack video={video} />
+            {projects.length && (
+              <Project items={projects} index={index} direction={direction} />
+            )}
+
+            {projects.length && (
+              <ItemControl length={projects.length} name={'Project'} />
+            )}
+          </section>
+        );
+      }}
+    </Consumer>
+  );
 }
-
-export default Projects;
